Region Rework 0.70.14.0

+++ regions/src/main/java/prplegoo/regions/buildings/supply/ROOM_LOGISTICS.java
@@ -37,7 +37,7 @@ public final class ROOM_LOGISTICS extends RoomBlueprintIns<SupplyInstance> imple
 	}
 
 	@Override
-	protected void update(float ds) {
+	protected void update(double ds) {
 		
 		
 	}
